+++ trionyx/trionyx/static/js/trionyx.js
@@ -232,6 +232,76 @@ function initGlobalSearch(searchUrl) {
     });
 }
 
+/* Tasks */
+function initTrionyxTasks(taskUrl) {
+    return new Vue({
+        el: '#trionyx-app-tasks',
+        delimiters: ['[[', ']]'],
+        data: {
+            taskUrl: taskUrl,
+            ajaxCall: null,
+
+            tasks: [],
+        },
+        computed: {
+            openTasks: function () {
+                return this.tasks.reduce(function (value, task) {
+                    if (task.status === 10 || task.status === 20 || task.status === 30 || task.status === 40) {
+                        return value + 1;
+                    }
+                    return value;
+                }, 0);
+            }
+        },
+        methods: {
+            getLabelClass(task) {
+                if (task.status === 10) {
+                    return 'label-warning';
+                } else if (task.status === 50) {
+                    return 'label-success';
+                } else if (task.status === 99) {
+                    return 'label-danger';
+                }
+                return 'label-info';
+            },
+            load: function() {
+                var self = this;
+
+                $.ajax({
+                    type: 'GET',
+                    url: this.taskUrl,
+                }).done(function(response) {
+                    if (response.status !== 'success') {
+                        return;
+                    }
+
+                    self.tasks = response.data;
+                    var timeout = 60;
+
+                    var runningTasks = self.tasks.reduce(function (value, task) {
+                        if (task.status === 20 || task.status === 30 || task.status === 40) {
+                            return value + 1;
+                        }
+
+                        return value;
+                    }, 0);
+
+                    if (runningTasks > 0) {
+                        timeout = 1;
+                    }
+
+                    setTimeout(function () {
+                        self.load();
+                    }, 1000 * timeout);
+                });
+            }
+        },
+        created: function() {
+            this.load();
+        },
+    });
+}
+
 /* Dialog */
 function openDialog(url, options) {
     new TrionyxDialog(url, options);